+++ src/error/compaction.rs
@@ -1,4 +1,4 @@
-use core::fmt;
+use core::{error::Error, fmt};
 
 /// Errors occurring while compacting a set of cell indices.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -21,9 +21,8 @@ impl fmt::Display for CompactionError {
     }
 }
 
-#[cfg(feature = "std")]
-impl std::error::Error for CompactionError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+impl Error for CompactionError {
+    fn source(&self) -> Option<&(dyn Error + 'static)> {
         None
     }
 }