+++ src/error/hex_grid.rs
@@ -1,4 +1,4 @@
-use core::fmt;
+use core::{error::Error, fmt};
 
 /// Errors related to the `IJK` coordinate system and its variants (e.g.
 /// [`LocalIJ`](crate::LocalIJ)).
@@ -20,9 +20,8 @@ impl fmt::Display for HexGridError {
     }
 }
 
-#[cfg(feature = "std")]
-impl std::error::Error for HexGridError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+impl Error for HexGridError {
+    fn source(&self) -> Option<&(dyn Error + 'static)> {
         None
     }
 }